added ClearCache.Mode with new verbose and simulate options

+++ FileHelper.cs
@@ -8,11 +8,11 @@ namespace SubTubular
     internal static class FileHelper
     {
         internal static IEnumerable<string> DeleteFiles(string directory, string searchPattern = "*",
-            ushort? notAccessedForDays = null)
+            ushort? notAccessedForDays = null, bool simulate = false)
         {
             foreach (var file in GetFiles(directory, searchPattern, notAccessedForDays))
             {
-                file.Delete();
+                if (!simulate) file.Delete();
                 yield return file.Name;
             }
         }
